requirements.txt
Flask
requests
gunicorn
pycryptodome
psycopg2
dotenv
urllib3
aiohttp
googleapis-common-protos
pycryptodome
protobuf